+++ models/trade/futureOrderRequest.go
@@ -3,7 +3,7 @@ package trade
 // FutureOrderRequest represents the request to set a future order
 type FutureOrderRequest struct {
 	CryptoID    uint    `json:"crypto_id,omitempty"`
-	Amount      float64 `json:"amount,omitempty"`
+	Amount      int 	`json:"amount,omitempty"`
 	TargetPrice int     `json:"target_price,omitempty"`
 	StopLoss    int     `json:"stop_loss,omitempty"`
 	TakeProfit  int     `json:"take_profit,omitempty"`
